Добавил абстрактный класс и вынес в него общие методы. Добавил наследование простым компонентам
