+++ src/view/user-profile-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view.js';
 
 const createUserProfileTemplate = () => (
   `<section class="header__profile profile">
@@ -7,22 +7,8 @@ const createUserProfileTemplate = () => (
   </section>`
 );
 
-export default class UserProfileView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class UserProfileView extends AbstractView {
   get template() {
     return createUserProfileTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
